Enable parameterized protocols by default, and add new staging flag for parameterized existentials
