+++ test/type/parameterized_protocol.swift
@@ -1,6 +1,6 @@
-// RUN: %target-typecheck-verify-swift -requirement-machine-protocol-signatures=on -requirement-machine-inferred-signatures=on -enable-parameterized-protocol-types -disable-availability-checking
+// RUN: %target-typecheck-verify-swift -requirement-machine-protocol-signatures=on -requirement-machine-inferred-signatures=on -disable-availability-checking
 
-// RUN: not %target-swift-frontend -typecheck %s -debug-generic-signatures -requirement-machine-protocol-signatures=on -enable-parameterized-protocol-types -requirement-machine-inferred-signatures=on -disable-availability-checking 2>&1 | %FileCheck %s
+// RUN: not %target-swift-frontend -typecheck %s -debug-generic-signatures -requirement-machine-protocol-signatures=on -requirement-machine-inferred-signatures=on -disable-availability-checking 2>&1 | %FileCheck %s
 
 
 /// Test some invalid syntax first
@@ -189,11 +189,16 @@ func returnsSequenceOfInt1() -> Sequence<Int> {}
 // expected-error@-1 {{protocol type with type arguments can only be used as a generic constraint}}
 
 func takesSequenceOfInt2(_: any Sequence<Int>) {}
+// expected-error@-1 {{protocol type with type arguments can only be used as a generic constraint}}
 
 func returnsSequenceOfInt2() -> any Sequence<Int> {}
+// expected-error@-1 {{protocol type with type arguments can only be used as a generic constraint}}
 
 func typeExpr() {
   _ = Sequence<Int>.self
+  // expected-error@-1 {{protocol type with type arguments can only be used as a generic constraint}}
+
+  _ = any Sequence<Int>.self
   // expected-error@-1 {{protocol type with type arguments can only be used as a generic constraint}}
 }
 
@@ -203,56 +208,3 @@ protocol SomeProto {}
 
 func protocolCompositionNotSupported(_: SomeProto & Sequence<Int>) {}
 // expected-error@-1 {{non-protocol, non-class type 'Sequence<Int>' cannot be used within a protocol-constrained type}}
-
-/// More regression tests
-
-protocol DoubleWide<X, Y> {
-  associatedtype X
-  associatedtype Y
-
-  var x: X { get }
-  var y: Y { get }
-}
-
-extension Int: DoubleWide {
-  typealias X = Int
-  typealias Y = Int
-
-  var x: X { 0 }
-  var y: X { 0 }
-}
-
-struct Collapse<T: DoubleWide>: DoubleWide {
-  typealias X = T
-  typealias Y = T
-
-  var x: X
-  var y: X { self.x }
-}
-
-func test() -> any DoubleWide<some DoubleWide<Int, Int>, some DoubleWide<Int, Int>> { return Collapse<Int>(x: 42) }
-
-func diagonalizeAny(_ x: any Sequence<Int>) -> any Sequence<(Int, Int)> {
-  return x.map { ($0, $0) }
-}
-
-func erase<T>(_ x: ConcreteSequence<T>) -> any Sequence<T> {
-  return x as any Sequence<T>
-}
-
-protocol Sponge<A, B> {
-  associatedtype A
-  associatedtype B
-}
-
-func saturation(_ dry: any Sponge, _ wet: any Sponge<Int, Int>) {
-  _ = dry as any Sponge<Int, Int>
-  // expected-error@-1 {{'any Sponge' is not convertible to 'any Sponge<Int, Int>'}}
-  // expected-note@-2 {{did you mean to use 'as!' to force downcast?}}
-  _ = dry as any Sponge
-
-  _ = wet as any Sponge<Int, Int> // Ok
-  _ = wet as any Sponge // Ok
-  _ = wet as any Sponge<String, String> // expected-error {{'any Sponge<Int, Int>' is not convertible to 'any Sponge<String, String>'}}
-  // expected-note@-1 {{did you mean to use 'as!' to force downcast?}}
-}
